fixed Toro summon issue
Now when we spawn a new Toro, it will be next to the player, and if we call summon and there is an existing Toro, it will move to the player. But if we run a program and a Toro exists, it will start the program from where it is standing.